Changing UI present and delete old code
Changing Gold changing UI and logic behind.
Changing Tier unlock value update.
Delete not used fields.
