+++ Assets/Code/ResourceMarket/MVC/Data/MarketDataConfig.cs
@@ -22,12 +22,4 @@ namespace ResourceMarket
         public float MarketCoef => _marketCoef;
         public float MarketRestoreValueDelay => _marketRestoreValueDelay;
     }
-
-    [System.Serializable]
-    public sealed class MarketTierData
-    {
-        [field: SerializeField] public int TierOneUnlockValue { get; private set; } = 0;
-        [field: SerializeField] public int TierTwoUnlockValue { get; private set; } = 3;
-        [field: SerializeField] public int TierThreeUnlockValue { get; private set; } = 6;
-    }
 }
