+++ Assets/Code/ResourceMarket/MVC/Data/MarketItemData.cs
@@ -15,7 +15,6 @@ namespace ResourceMarket
         [field: SerializeField] public int InitialAmount { get; private set; } = 100;
         [field: SerializeField] public int ExchangeAmount { get; private set; } = 10;
         [field: SerializeField] public int ExchangeRate { get; private set; } = 100;
-        [field: SerializeField] public int MinExchange { get; private set; } = 50;
         [field: SerializeField] public float ExchangeCoef { get; private set; } = 0.45f;
 
         [field: SerializeField] public string Name { get; private set; }
