+++ Assets/Code/ResourceMarket/MVC/Data/MarketTierData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1881224139d8ec54c981df9a86635b80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
